+++ README.md
@@ -37,7 +37,7 @@ Database drivers run migrations. [Add a new database?](database/driver.go)
   * [CockroachDB](database/cockroachdb)
   * [ClickHouse](database/clickhouse)
   * [Firebird](database/firebird) ([todo #49](https://github.com/golang-migrate/migrate/issues/49))
-  * [MS SQL Server](database/mssql)
+  * [MS SQL Server](database/sqlserver)
 
 ### Database URLs
