+++ database/sqlserver/sqlserver.go
@@ -1,4 +1,4 @@
-package mssql
+package sqlserver
 
 import (
 	"context"
@@ -15,8 +15,7 @@ import (
 )
 
 func init() {
-	db := MSSQL{}
-	database.Register("sqlserver", &db)
+	database.Register("sqlserver", &SQLServer{})
 }
 
 // DefaultMigrationsTable is the name of the migrations table in the database
@@ -44,7 +43,7 @@ type Config struct {
 }
 
 // MSSQL connection
-type MSSQL struct {
+type SQLServer struct {
 	// Locking and unlocking need to use the same connection
 	conn     *sql.Conn
 	db       *sql.DB
@@ -100,7 +99,7 @@ func WithInstance(instance *sql.DB, config *Config) (database.Driver, error) {
 		return nil, err
 	}
 
-	ss := &MSSQL{
+	ss := &SQLServer{
 		conn:   conn,
 		db:     instance,
 		config: config,
@@ -114,7 +113,7 @@ func WithInstance(instance *sql.DB, config *Config) (database.Driver, error) {
 }
 
 // Open a connection to the database
-func (ss *MSSQL) Open(url string) (database.Driver, error) {
+func (ss *SQLServer) Open(url string) (database.Driver, error) {
 	purl, err := nurl.Parse(url)
 	if err != nil {
 		return nil, err
@@ -140,7 +139,7 @@ func (ss *MSSQL) Open(url string) (database.Driver, error) {
 }
 
 // Close the database connection
-func (ss *MSSQL) Close() error {
+func (ss *SQLServer) Close() error {
 	connErr := ss.conn.Close()
 	dbErr := ss.db.Close()
 	if connErr != nil || dbErr != nil {
@@ -150,7 +149,7 @@ func (ss *MSSQL) Close() error {
 }
 
 // Lock creates an advisory local on the database to prevent multiple migrations from running at the same time.
-func (ss *MSSQL) Lock() error {
+func (ss *SQLServer) Lock() error {
 	if ss.isLocked {
 		return database.ErrLocked
 	}
@@ -177,7 +176,7 @@ func (ss *MSSQL) Lock() error {
 }
 
 // Unlock froms the migration lock from the database
-func (ss *MSSQL) Unlock() error {
+func (ss *SQLServer) Unlock() error {
 	if !ss.isLocked {
 		return nil
 	}
@@ -198,7 +197,7 @@ func (ss *MSSQL) Unlock() error {
 }
 
 // Run the migrations for the database
-func (ss *MSSQL) Run(migration io.Reader) error {
+func (ss *SQLServer) Run(migration io.Reader) error {
 	migr, err := ioutil.ReadAll(migration)
 	if err != nil {
 		return err
@@ -221,7 +220,7 @@ func (ss *MSSQL) Run(migration io.Reader) error {
 }
 
 // SetVersion for the current database
-func (ss *MSSQL) SetVersion(version int, dirty bool) error {
+func (ss *SQLServer) SetVersion(version int, dirty bool) error {
 
 	tx, err := ss.conn.BeginTx(context.Background(), &sql.TxOptions{})
 	if err != nil {
@@ -258,7 +257,7 @@ func (ss *MSSQL) SetVersion(version int, dirty bool) error {
 }
 
 // Version of the current database state
-func (ss *MSSQL) Version() (version int, dirty bool, err error) {
+func (ss *SQLServer) Version() (version int, dirty bool, err error) {
 	query := `SELECT TOP 1 version, dirty FROM "` + ss.config.MigrationsTable + `"`
 	err = ss.conn.QueryRowContext(context.Background(), query).Scan(&version, &dirty)
 	switch {
@@ -275,7 +274,7 @@ func (ss *MSSQL) Version() (version int, dirty bool, err error) {
 }
 
 // Drop all tables from the database.
-func (ss *MSSQL) Drop() error {
+func (ss *SQLServer) Drop() error {
 
 	// drop all referential integrity constraints
 	query := `
@@ -309,7 +308,7 @@ func (ss *MSSQL) Drop() error {
 	return nil
 }
 
-func (ss *MSSQL) ensureVersionTable() (err error) {
+func (ss *SQLServer) ensureVersionTable() (err error) {
 	if err = ss.Lock(); err != nil {
 		return err
 	}
