--- internal/cli/build_mssql.go
@@ -1,7 +1,0 @@
-// +build mssql
-
-package cli
-
-import (
-	_ "github.com/golang-migrate/migrate/v4/database/mssql"
-)
