+++ internal/cli/build_sqlserver.go
@@ -1,0 +1,7 @@
+// +build sqlserver
+
+package cli
+
+import (
+	_ "github.com/golang-migrate/migrate/v4/database/sqlserver"
+)
